refactor(builder): move undo/redo to the dock and center the header title
- move the Undo/Redo buttons (and their Mod+Z / Mod+Shift+Z / Ctrl+Y hotkeys) from the
  header into the bottom dock, grouped before the zoom controls
- restructure the header into equal-width flex-1 side groups so the centered title +
  actions are truly centered instead of skewing left under the wider Download button

Verified in-browser: header center offset is 0px; dock undo reverts an edit.
